+++ km_api/know_me/conftest.py
@@ -49,3 +49,11 @@ def km_user_factory(db):
         The factory class used to create test ``KMUser`` instances.
     """
     return factories.KMUserFactory
+
+
+@pytest.fixture
+def legacy_user_factory(db):
+    """
+    Fixture to get the factory used to create legacy users.
+    """
+    return factories.LegacyUserFactory